Portals kinda working
Teleporting from portal A to B in my scene works 80% of the time but Teleporting from B to A doesnt work at all. The cameras are still doing the weird glitch where when you press play they will move out position I think its something to do with the calculation to see the distance between portals or something. Texturing on portals works just cameras on the portals are the problem rn. See if you can fix this plez.