improved picking up objects
now less likely to go through walls (if rigidbody has continuous speculative detection and no interpolation), and smoothly moves position and rotation to always center infront of player. Wall detection works by parenting and de-parenting the object to and from the player in order to keep allowing the portals to work with it, otherwise I wouldnt bother with parenting at all. parenting is essential for the smooth passthrough and game integrity